# Arguments rewrite