… .travis.yml
… devito/core/autotuning.py
… devito/dimension.py
… devito/dle/backends/basic.py
… devito/foreign/__init__.py
… devito/function.py
… devito/ir/iet/nodes.py
… devito/ir/iet/visitors.py
… devito/ir/support/stencil.py
… devito/operator.py
… devito/tools.py
… devito/types.py
… devito/yask/operator.py
… examples/cfd/01_convection.ipynb
… examples/cfd/01_convection_revisited.ipynb
… examples/checkpointing/checkpoint.py
… tests/test_adjointJ.py
… tests/test_checkpointing.py
… tests/test_operator.py