Inviting review on this even though it doesn't pass all the tests yet. 

Edit: This PR now passes all tests