@navjotk 2017-12-30
> I also still don't understand the reluctance towards "all parameters have providers". It seems to work pretty smoothly?

It was working smoothly without it as well, except for that one line in the yask specific argument handling that I wasn't willing to touch in this PR to minimise the footprint of this PR. If we can do without providers on all parameters, that would be a simpler and hence better solution no?


> There's still some clean up to do, as the TODOs point out. Is this the next step? or is this stuff for the longer term?

Maybe not the immediate next step but soon. 

> For example, the handling of the dle_arguments.

I am aware that can be simplified further but requires a small change in the DLE classes regarding autotuning, which also I left for the future in order to keep this PR small. 

> I also assume that you measure the Python overhead of this new solution and it's basically negligible? ie, from calling op.apply() to the actual execution of the cfunction, the overhead is negligible?

I will measure it and post it here. I assume you mean in comparison to master? In absolute terms I won't be surprised if it is not negligible but it should be the same or less than in master. 

> Does this PR fix #416 , or any other standing issues?

It doesn't yet, but brings us closer. This is a good way along to fixing most arguments issues though, in my opinion.