@FabioLuporini 2018-01-02
I see you changed the a few things in the autotuner. Could you check that ``acoustic_example`` and/or ``tti_example`` still take up about the same time as on ``master`` when ran with ``DEVITO_AUTOTUNING=aggressive`` ? When using ``DEVITO_LOGGING=DEBUG``, you should see lots of AT prints.

Then I think we're ready to merge