@navjotk 2018-01-03
Timings for the following code:
```
from examples.seismic.acoustic.acoustic_example import acoustic_setup
solver = acoustic_setup()
solver.op_fwd()
%timeit solver.op_fwd().arguments()
```
Master:
633 µs ± 34.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

This branch:
52.7 ms ± 2.95 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)