@FabioLuporini 2018-01-03
> Could you check that acoustic_example and/or tti_example still take up about the same time as on master when ran with DEVITO_AUTOTUNING=aggressive ? When using DEVITO_LOGGING=DEBUG, you should see lots of AT prints.

any news on this?

> Master:
> 633 µs ± 34.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
> 
> This branch:
> 52.7 ms ± 2.95 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

The slow down is really, really high... is it worth investigating why? Overall, it appears to be negligible though... If TTI doesn't get incredibly slow, we might accept this. @mlange05 ?