@navjotk 2018-01-03
> Could you check that acoustic_example and/or tti_example still take up about the same time as on master when ran with DEVITO_AUTOTUNING=aggressive ? When using DEVITO_LOGGING=DEBUG, you should see lots of AT prints.

I don't see `DEVITO_AUTOTUNING` in our configuration [dict](https://github.com/opesci/devito/blob/master/devito/parameters.py#L95). Not sure if this even exists in master?