@navjotk 2018-01-03
Running the examples takes about the same time though. 
Using:
```
from examples.seismic.acoustic.acoustic_example import run as acoustic_run
from examples.seismic.tti.tti_example import run as tti_run
%timeit acoustic_run()
%timeit tti_run()
```
On master:
Acoustic: 10 s ± 457 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
tti: 8.89 s ± 395 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

This branch:
acoustic: 9.51 s ± 291 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
tti: 9.19 s ± 64.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

The difference seems to be within the margin of measurement error in both cases.