@FabioLuporini 2018-01-03
DEVITO_AUTOTUNING is a perfectly legal Devito option. Have you tried using it? 
Its implementation is here: https://github.com/opesci/devito/blob/master/devito/core/__init__.py

Can you try running the examples above with AT, ie passing ``-a``? Do they take as long as in master? I'm interested in the number of AT prints and how long each AT run takes (DEVITO_LOGGING=DEBUG will help)