@navjotk 2018-01-03
I think it is very bad design to have configuration options spread across the code base.