@navjotk 2018-01-03
This fails: 
```
from devito import configuration
configuration['autotuning'] = 'aggressive'
```