@FabioLuporini 2018-01-03
> I think it is very bad design to have configuration options spread across the code base.

I disagree. That `autotuning` is a core-specific parameter. 

Also, potentially bad design choices should be detected at PR review time



> This fails:
> 
> from devito import configuration
> configuration['autotuning'] = 'aggressive'

This will not:
configuration.core['autotuning'] = 'aggressive'