@navjotk 2018-01-03
> That autotuning is a core-specific parameter.

Doesn't explain why it is in `devito/__init__.py`. Not the first place one would expect configuration options to be. There should either be a backend specific dict next to the main one, or this should be in a backend specific configuration module. 

> Also, potentially bad design choices should be detected at PR review time

Maybe the reviewer should have applied stricter review standards. Or let the PR be open long enough for someone else to take a look. Like in this PR, for example.