@FabioLuporini 2018-01-03
> Doesn't explain why it is in devito/__init__.py. Not the first place one would expect configuration options to be. There should either be a backend specific dict next to the main one, or this should be in a backend specific configuration module.

No. It's not there, it's in ``devito/core/__init__.py``. And I think you haven't completely understood how the configuration mechanism works. For example:

> There should either be a backend specific dict next to the main one, or this should be in a backend specific configuration module.

That's how it is.

> Yes I agree that the review standards applied to that PR don't match up to those being applied to this PR. Maybe if it remained open for longer, other people might have had a chance to take a look.

No, the team is small and PR reviewing has to be relatively quick. If it weren't quick, there would have been periods in which I would have ended up stacking up tons of PRs. BTW, most PRs have been open for several days; only the very small ones for 1-2.