@navjotk 2018-01-03
> I think you haven't completely understood how the configuration mechanism works

Maybe that's a sign of missing and required documentation.

> configuration.core['autotuning'] = 'aggressive'

This doesn't seem to change anything in the execution of the code I posted above. No autotuning is being performed.