@FabioLuporini 2018-01-04
As I said, it'd be very useful to verify the correct functioning of autotuning after your changes to ``autotuning.py``. This is probably the last standing item