@navjotk 2018-01-04
As I said above, the code:
```
from examples.seismic.acoustic.acoustic_example import run as acoustic_run
from examples.seismic.tti.tti_example import run as tti_run
from devito import configuration
configuration.core['autotuning'] = 'aggressive'
configuration['log_level'] = 'debug'
%timeit acoustic_run()
%timeit tti_run()
```
does not show any log messages related to autotuning, whether on master or this branch.