@FabioLuporini 2018-01-04
have you considered passing ``autotune=True``, or running from the command line with ``-a`` ?