@navjotk 2018-01-05
The change that I had made in the autotuner was actually causing other issues that made `tti_example.py` segfault with autotuning enabled so I reverted to using the backdoor arguments that you were using. I guess no performance evaluation necessary now.